backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pdfplumber==0.10.3
anthropic==0.39.0
python-multipart==0.0.6
openpyxl==3.1.2
python-jose[cryptography]==3.3.0
